correct bal td compile warning (#4440)
* correct bal td compile warning

* fix introduced issues

+++ groups/bal/balm/balm_metrics.t.cpp
@@ -187,8 +187,8 @@ bool within(double         value,
     // within the specified 'windowMs' (milliseconds) of the specified
     // 'expectedS' (seconds).
 {
-    double expected = expectedS * scale;
-    double window   = windowMs *1000 * scale;
+    double expected = expectedS * static_cast<double>(scale);
+    double window   = windowMs * 1000 * static_cast<double>(scale);
 
     return ((expected - window) < value)
         && ((expected + window) > value);
